refactor: app implementation decoupling
