+++ app/constants.py
@@ -1,12 +1,1 @@
 API_VERSION = "v0.5.0beta"
-
-available_app_ids = {
-    "btc-rpc-explorer",
-    "rtl",
-    # Specter is deactivated for now because it uses its own selfsigned HTTPS cert that makes trouble in chome on last test
-    # "specter",
-    "btcpayserver",
-    "lnbits",
-    "mempool",
-    "thunderhub",
-}
