+++ app/repositories/apps_impl/apps_base.py
@@ -1,0 +1,15 @@
+class AppsBase:
+    async def get_app_status_single(self, app_id: str):
+        raise NotImplementedError()
+
+    async def get_app_status(self):
+        raise NotImplementedError()
+
+    async def get_app_status_sub(self):
+        raise NotImplementedError()
+
+    async def install_app_sub(self, app_id: str):
+        raise NotImplementedError()
+
+    async def uninstall_app_sub(self, app_id: str, delete_data: bool):
+        raise NotImplementedError()
