+++ app/repositories/apps_impl/native_python.py
@@ -1,0 +1,5 @@
+from app.repositories.apps_impl.apps_base import AppsBase
+
+
+class NativePythonAppsImpl(AppsBase):
+    pass
